refactor(core-weave): extract legacy page renderers
