+++ documentation/notes/wd.codebase-overview.md
@@ -33,7 +33,7 @@ Core owns portable Semantic Flow behavior: request/result types, RDF/Turtle help
 - `src/core/mesh`, `src/core/knop`, `src/core/integrate`, `src/core/payload`, and `src/core/extract` own operation-specific pure planning.
 - `src/core/weave` owns shared weave/version/generate contracts and planner helpers.
 - `src/core/weave/weave.ts` is still the major pressure point. It remains the public façade for core weave imports, but it is being decomposed under [[wa.task.2026.2026-05-21_0849_careful-extraction-refactor]].
-- Already extracted core weave helpers include `errors.ts`, `version_plan.ts`, `mesh_support_pages.ts`, `requests.ts`, `source_models.ts`, `candidates.ts`, `planning_models.ts`, `progression_models.ts`, `slices.ts`, `rdf_helpers.ts`, `turtle_blocks.ts`, `artifact_history_queries.ts`, `artifact_manifestation_paths.ts`, `slice_classification.ts`, `payload_version_layout.ts`, `payload_overwrite.ts`, `payload_renderers.ts`, `mesh_inventory_renderers.ts`, `knop_inventory_renderers.ts`, `knop_support_renderers.ts`, `shape_assertions.ts`, `source_locator_assertions.ts`, `source_locator_renderers.ts`, `support_history_renderers.ts`, `working_file_paths.ts`, `reference_catalog_links.ts`, `resource_page_builders.ts`, `resource_page_models.ts`, `resource_page_history_groups.ts`, `resource_page_policy.ts`, `resource_page_reference_links.ts`, `naming_policy.ts`, and `support_history_policy.ts`.
+- Already extracted core weave helpers include `errors.ts`, `version_plan.ts`, `mesh_support_pages.ts`, `requests.ts`, `source_models.ts`, `candidates.ts`, `planning_models.ts`, `progression_models.ts`, `slices.ts`, `rdf_helpers.ts`, `turtle_blocks.ts`, `artifact_history_queries.ts`, `artifact_manifestation_paths.ts`, `slice_classification.ts`, `payload_version_layout.ts`, `payload_overwrite.ts`, `payload_renderers.ts`, `mesh_inventory_renderers.ts`, `knop_inventory_renderers.ts`, `legacy_page_renderers.ts`, `knop_support_renderers.ts`, `shape_assertions.ts`, `source_locator_assertions.ts`, `source_locator_renderers.ts`, `support_history_renderers.ts`, `working_file_paths.ts`, `reference_catalog_links.ts`, `resource_page_builders.ts`, `resource_page_models.ts`, `resource_page_history_groups.ts`, `resource_page_policy.ts`, `resource_page_reference_links.ts`, `naming_policy.ts`, and `support_history_policy.ts`.
 
 ## Runtime Layer
 
@@ -81,7 +81,8 @@ Runtime owns local execution against a workspace: filesystem reads/writes, git-a
 - Completed core weave slice-classification extraction slice: [[wa.completed.2026.2026-05-22_1424-core-weave-slice-classification-extraction]].
 - Completed core weave payload renderer extraction slice: [[wa.completed.2026.2026-05-22_2252-payload-render-helpers]].
 - Completed core weave Knop support preservation extraction slice: [[wa.completed.2026.2026-05-22_2117-core-weave-knop-support-render-preservation-extraction]].
-- Implemented core weave mesh inventory renderer extraction slice: [[wa.task.2026.2026-05-22_2139-core-weave-mesh-inventory-renderer-extraction]].
-- Latest core weave KnopInventory renderer extraction slice: [[wa.task.2026.2026-05-22_2206-core-weave-knop-inventory-renderer-extraction]].
+- Completed core weave mesh inventory renderer extraction slice: [[wa.completed.2026.2026-05-22_2139-core-weave-mesh-inventory-renderer-extraction]].
+- Completed core weave KnopInventory renderer extraction slice: [[wa.completed.2026.2026-05-22_2206-core-weave-knop-inventory-renderer-extraction]].
+- Latest core weave legacy HTML page renderer extraction slice: [[wa.task.2026.2026-05-22_2222-core-weave-html-page-renderer-extraction]].
 - Runtime weave decomposition: [[wa.task.2026.2026-05-21_1035-runtime-weave-module-decomposition]].
 - Runtime ResourcePage generation decomposition is complete: [[wa.completed.2026.2026-05-21_1036-runtime-resource-page-generation-decomposition]].
